fix: refactor base and stories
