--- lib/components/Modal/bottom/modal-bottom.scss
@@ -1,22 +1,0 @@
-.modal-bottom {
-    &__container {
-        position: absolute;
-        bottom: 0;
-        width: 100%;
-        max-height: 90%;
-        min-height: var(--temp-static-spacing-289);
-        overflow: hidden;
-        display: flex;
-        align-items: stretch;
-        justify-content: space-evenly;
-        flex-direction: column;
-        background-color: var(--component-modal-bg);
-        border-top-left-radius: var(--core-borderRadius-800);
-        border-top-right-radius: var(--core-borderRadius-800);
-        opacity: var(--core-opacity-50);
-        transform: translateY(100%);
-        pointer-events: none;
-        transition: all var(--core-motion-duration-300)
-            var(--motion-easing-linear); 
-    }
-}
