+++ lib/components/Modal/index.tsx
@@ -1,19 +1,19 @@
 import React, { ReactNode } from "react";
-import { ModalHeader } from "./modal-header";
-import { ModalBody } from "./modal-body";
-import { ModalBase } from "./base";
+import { ModalHeader } from "./base/modal-header";
+import { ModalBody } from "./base/modal-body";
+import { ModalBase } from "./base/base";
 
 export interface ModalType {
     Header: typeof ModalHeader;
     Body: typeof ModalBody;
-    Base: typeof ModalBase;
+    Frame: typeof ModalBase;
 }
 
 export const Modal: ModalType = ({ children }: { children: ReactNode }) => {
     return <>{children}</>;
 };
 
-Modal.Base = ModalBase;
+Modal.Frame = ModalBase;
 Modal.Header = ModalHeader;
 Modal.Body = ModalBody;
