--- lib/components/Modal/overlay/modal-overlay.scss
@@ -1,14 +1,0 @@
-.modal-overlay {
-    &__container {
-        position: relative;
-        overflow: hidden;
-        display: flex;
-        border-radius: var(--core-borderRadius-1600);
-        width: 512px;
-        max-height: 720px;
-        justify-content: space-evenly;
-        flex-direction: column;
-        background-color: var(--component-modal-bg);
-        margin : 0 auto;
-    }
-}
